+++ packages/moonstone/ExpandableItem/tests/ExpandableItem-specs.js
@@ -3,35 +3,35 @@ import {ExpandableItemBase} from '../ExpandableItem';
 describe('ExpandableItem', () => {
 	describe('computed', () => {
 		describe('label', () => {
-			it('should use noneText when label is not set', function () {
+			test('should use noneText when label is not set', () => {
 				const expected = 'noneText';
 				const actual = ExpandableItemBase.computed.label({
 					noneText: 'noneText'
 				});
 
-				expect(actual).to.equal(expected);
+				expect(actual).toBe(expected);
 			});
 
-			it('should use label when set', function () {
+			test('should use label when set', () => {
 				const expected = 'label';
 				const actual = ExpandableItemBase.computed.label({
 					label: 'label',
 					noneText: 'noneText'
 				});
 
-				expect(actual).to.equal(expected);
+				expect(actual).toBe(expected);
 			});
 		});
 
 		describe('open', () => {
-			it('should be false when disabled', function () {
+			test('should be false when disabled', () => {
 				const expected = false;
 				const actual = ExpandableItemBase.computed.open({
 					disabled: true,
 					open: true
 				});
 
-				expect(actual).to.equal(expected);
+				expect(actual).toBe(expected);
 			});
 		});
 	});