+++ packages/moonstone/Picker/tests/Picker-specs.js
@@ -3,7 +3,7 @@ import {mount, shallow} from 'enzyme';
 import {Picker, PickerBase} from '../Picker';
 
 describe('Picker Specs', () => {
-	it('should render selected child wrapped with <PickerItem/>', function () {
+	test('should render selected child wrapped with <PickerItem/>', () => {
 		const picker = mount(
 			<Picker value={1}>
 				{[1, 2, 3, 4]}
@@ -13,24 +13,27 @@ describe('Picker Specs', () => {
 		const expected = '2';
 		const actual = picker.find('PickerItem').text();
 
-		expect(actual).to.equal(expected);
+		expect(actual).toBe(expected);
 	});
 
-	it('should set the max of <Picker> to be one less than the number of children', function () {
-		const picker = mount(
-			<Picker value={1}>
-				{[1, 2, 3, 4]}
-			</Picker>
-		);
+	test(
+		'should set the max of <Picker> to be one less than the number of children',
+		() => {
+			const picker = mount(
+				<Picker value={1}>
+					{[1, 2, 3, 4]}
+				</Picker>
+			);
 
-		const expected = 3;
-		const actual = picker.find('Picker').last().prop('max');
+			const expected = 3;
+			const actual = picker.find('Picker').last().prop('max');
 
-		expect(actual).to.equal(expected);
-	});
+			expect(actual).toBe(expected);
+		}
+	);
 
 
-	it('should be disabled when empty', function () {
+	test('should be disabled when empty', () => {
 		const picker = shallow(
 			<PickerBase>
 				{[]}
@@ -38,6 +41,6 @@ describe('Picker Specs', () => {
 		);
 
 		const actual = picker.find('SpottablePicker').last().prop('disabled');
-		expect(actual).to.be.true();
+		expect(actual).toBe(true);
 	});
 });