+++ packages/spotlight/src/tests/pointer-specs.js
@@ -18,11 +18,11 @@ describe('pointer', () => {
 	afterEach(reset);
 
 	describe('#getLastPointerPosition', () => {
-		it('should return an object with x and y keys', () => {
+		test('should return an object with x and y keys', () => {
 			const expected = ['x', 'y'];
 			const actual = Object.keys(getLastPointerPosition());
 
-			expect(actual).to.deep.equal(expected);
+			expect(actual).toEqual(expected);
 		});
 	});
 
@@ -36,62 +36,62 @@ describe('pointer', () => {
 			pointerShow: 888
 		};
 
-		before(() => addAll(keyMap));
-		after(() => removeAll(keyMap));
+		beforeAll(() => addAll(keyMap));
+		afterAll(() => removeAll(keyMap));
 
 		beforeEach(() => {
 			// establish a consistent pointer mode state for each test
 			setPointerMode(false);
 		});
 
-		it('should return true for pointer hide key events', () => {
+		test('should return true for pointer hide key events', () => {
 			const expected = true;
 			const actual = notifyKeyDown(keyMap.pointerHide);
 
-			expect(actual).to.equal(expected);
+			expect(actual).toBe(expected);
 		});
 
-		it('should return true for pointer show key events', () => {
+		test('should return true for pointer show key events', () => {
 			const expected = true;
 			const actual = notifyKeyDown(keyMap.pointerShow);
 
-			expect(actual).to.equal(expected);
+			expect(actual).toBe(expected);
 		});
 
-		it('should enable pointer mode for pointer show key events', () => {
+		test('should enable pointer mode for pointer show key events', () => {
 			notifyKeyDown(keyMap.pointerShow);
 
 			const expected = true;
 			const actual = getPointerMode();
 
-			expect(actual).to.equal(expected);
+			expect(actual).toBe(expected);
 		});
 
-		it('should disable pointer mode for pointer show key events', (done) => {
+		test('should disable pointer mode for pointer show key events', (done) => {
 			setPointerMode(true);
 			notifyKeyDown(keyMap.pointerHide, () => {
 				const expected = false;
 				const actual = getPointerMode();
 
-				expect(actual).to.equal(expected);
+				expect(actual).toBe(expected);
 
 				done();
 			});
 		});
 
-		it('should disable pointer mode for non-pointer key events', () => {
+		test('should disable pointer mode for non-pointer key events', () => {
 			setPointerMode(true);
 			notifyKeyDown(12);
 
 			const expected = false;
 			const actual = getPointerMode();
 
-			expect(actual).to.equal(expected);
+			expect(actual).toBe(expected);
 		});
 	});
 
 	describe('#notifyPointerMove', () => {
-		it('should update the pointer position if x changes', () => {
+		test('should update the pointer position if x changes', () => {
 			const x = 20;
 
 			notifyPointerMove(null, null, x, null);
@@ -99,10 +99,10 @@ describe('pointer', () => {
 			const expected = 20;
 			const actual = getLastPointerPosition().x;
 
-			expect(actual).to.equal(expected);
+			expect(actual).toBe(expected);
 		});
 
-		it('should update the pointer position if y changes', () => {
+		test('should update the pointer position if y changes', () => {
 			const y = 20;
 
 			notifyPointerMove(null, null, null, y);
@@ -110,68 +110,77 @@ describe('pointer', () => {
 			const expected = 20;
 			const actual = getLastPointerPosition().y;
 
-			expect(actual).to.equal(expected);
+			expect(actual).toBe(expected);
 		});
 
-		it('should enable pointer mode if the pointer positionchanges', () => {
+		test('should enable pointer mode if the pointer positionchanges', () => {
 			setPointerMode(false);
 			notifyPointerMove(null, null, 5, 5);
 
 			const expected = true;
 			const actual = getPointerMode();
 
-			expect(actual).to.equal(expected);
+			expect(actual).toBe(expected);
 		});
 
-		it('should return false if the pointer has not moved', () => {
+		test('should return false if the pointer has not moved', () => {
 			notifyPointerMove(null, null, 5, 5);
 
 			const expected = false;
 			const actual = notifyPointerMove(null, null, 5, 5);
 
-			expect(actual).to.equal(expected);
+			expect(actual).toBe(expected);
 		});
 
-		it('should return true if pointer mode was disabled', () => {
+		test('should return true if pointer mode was disabled', () => {
 			// change into pointer mode indicates a potential for change of focus
 			setPointerMode(false);
 
 			const expected = true;
 			const actual = notifyPointerMove(null, null, 5, 5);
 
-			expect(actual).to.equal(expected);
+			expect(actual).toBe(expected);
 		});
 
-		it('should return true if the pointer has moved and current is falsy', () => {
-			const expected = true;
-			const actual = notifyPointerMove(null, null, 5, 5);
+		test(
+			'should return true if the pointer has moved and current is falsy',
+			() => {
+				const expected = true;
+				const actual = notifyPointerMove(null, null, 5, 5);
 
-			expect(actual).to.equal(expected);
-		});
+				expect(actual).toBe(expected);
+			}
+		);
 
-		it('should return false if the pointer has moved and target is within current', () => {
-			const current = document.createElement('div');
-			const target = document.createElement('div');
-			current.appendChild(target);
+		test(
+			'should return false if the pointer has moved and target is within current',
+			() => {
+				const current = document.createElement('div');
+				const target = document.createElement('div');
+				current.appendChild(target);
 
-			notifyPointerMove(null, current, 5, 5);
+				notifyPointerMove(null, current, 5, 5);
 
-			const expected = false;
-			const actual = notifyPointerMove(current, target, 10, 10);
+				const expected = false;
+				const actual = notifyPointerMove(current, target, 10, 10);
 
-			expect(actual).to.equal(expected);
-		});
+				expect(actual).toBe(expected);
+			}
+		);
 
-		it('should return true if the pointer has moved and target is not within current', () => {
-			const current = document.createElement('div');
-			const target = document.createElement('div');
+		test(
+			'should return true if the pointer has moved and target is not within current',
+			() => {
+				const current = document.createElement('div');
+				const target = document.createElement('div');
 
-			notifyPointerMove(null, current, 5, 5);
+				notifyPointerMove(null, current, 5, 5);
 
-			const expected = true;
-			const actual = notifyPointerMove(current, target, 10, 10);
+				const expected = true;
+				const actual = notifyPointerMove(current, target, 10, 10);
 
-			expect(actual).to.equal(expected);
-		});
+				expect(actual).toBe(expected);
+			}
+		);
 	});
 });