+++ packages/ui/Marquee/MarqueeDecorator.js
@@ -280,8 +280,8 @@ const MarqueeDecorator = hoc(defaultConfig, (config, Wrapped) => {
 			this.sync = false;
 			this.forceRestartMarquee = false;
 			this.timerState = TimerState.CLEAR;
-
-			this.invalidateMetrics();
+			this.distance = null;
+			this.contentFits = false;
 		}
 
 		componentWillMount () {
@@ -443,6 +443,10 @@ const MarqueeDecorator = hoc(defaultConfig, (config, Wrapped) => {
 			this.distance = null;
 			// Assume the marquee does not fit until calculations show otherwise
 			this.contentFits = false;
+
+			this.setState(state => {
+				return state.overflow === 'ellipsis' ? null : {overflow: 'ellipsis'};
+			});
 		}
 
 		/*