+++ packages/ui/CHANGELOG.md
@@ -8,6 +8,13 @@ The following is a curated list of changes in the Enact ui module, newest change
 
 - `ui/Toggleable` props `toggleOffAriaLabel` and `toggleOnAriaLabel` to configure the accessibility string in the toggleable.
 
+## [2.2.8] - 2018-12-06
+
+### Fixed
+
+- `ui/Marquee` to display an ellipsis when changing to text that no longer fits within its bounds
+- `ui/VirtualList`, `ui/VirtualGridList`, and `ui/Scroller` to debounce `onScrollStop` events for non-animated scrolls
+
 ## [2.2.7] - 2018-11-21
 
 ### Fixed